feat: export GEDCOM 5.5.1 BLOB decoder as go-glx DecodeGEDCOMBlob
Move decodeGEDCOMBlob from the CLI (glx/media_copy.go) into go-glx as the
exported DecodeGEDCOMBlob, together with its sentinel errors
(ErrEmptyBlobData, ErrInvalidBlobLength, ErrInvalidBlobChar). The CLI's
copyMediaFiles becomes a caller; library consumers receiving
MediaFileSource.BlobData no longer need to reimplement the decoding.

The full-group invalid-character error now wraps ErrInvalidBlobChar like
the trailing-group path already did, so all invalid-character failures
are detectable via errors.Is. Tests move to go-glx and gain errors.Is
assertions plus a known-value (0xFF) decode case.

Prerequisite for library-level GEDZIP import (#1141).

Closes #1140
