+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 
+- **`go-glx`: exported GEDCOM 5.5.1 BLOB decoder (`DecodeGEDCOMBlob`)** — The BLOB decoder previously lived unexported in the CLI (`decodeGEDCOMBlob` in `glx/media_copy.go`), so library consumers receiving `MediaFileSource.BlobData` had to reimplement the 5.5.1 decoding — including validation and the trailing-group edge cases — themselves. The function now lives in `go-glx` as `DecodeGEDCOMBlob`, together with its sentinel errors `ErrEmptyBlobData`, `ErrInvalidBlobLength`, and `ErrInvalidBlobChar` (moved out of the CLI package), and the CLI is a caller. Pure function, no I/O. The full-group invalid-character error now wraps `ErrInvalidBlobChar` like the trailing-group path already did, so all invalid-character failures are detectable via `errors.Is` (the message gains the "must be in range '.' to 'm'" hint). Prerequisite for library-level GEDZIP import (#1141). (#1140)
+
 - **`glx publish` — static HTML site generator** — Exports a GLX archive as a self-contained, browsable website for sharing with non-technical family. Generates a landing page with archive stats and an alphabetical person list, a profile page per person (vital facts, hierarchical birth/death places, a life timeline including family events, linked parents/spouses/children/siblings, supporting sources with citation locators and quotes, notes, and a media gallery), source and place indexes (places with coordinates link to OpenStreetMap), and a client-side search page. Output is a plain directory of HTML/CSS/JS — no server, database, or build tooling — that works when opened directly via `file://` (the search index loads as a `<script>`, not `fetch`) or hosted anywhere (GitHub Pages, S3, Netlify). Templates and assets are embedded in the binary. Styling is responsive with a light/dark toggle and print-friendly profile pages, and all archive content is HTML-escaped. Flags: `--output`/`-o` (default `./site`), `--embed-media` to inline images as base64, `--living` to redact likely-living people before publishing (reusing the `--privatize-living` redaction), and `--title`. Implements the first phase of the read-only web viewer; interactive SVG pedigree/descendancy charts, a visual timeline view, embedded maps, and theming config are tracked as follow-ups. (#91) Renamed from `glx view` before release to avoid ambiguity with `glx serve` and to name the artifact it produces; the `--serve`/`--port` preview flags were dropped in the same change, since `glx serve` now owns local preview (#970).
 - **Git Workflow Guide (`docs/guides/git-workflow.md`)** — New user guide documenting how to actually use Git workflows for genealogy research, filling the gap behind the "Git-native" headline feature: setting up an archive repository (`git init` after `glx init`, line-ending pinning, private-by-default remotes for living-person data), the edit→`glx validate`→commit research cycle with a pre-commit validation hook, branch-based research (hypothesis branches carrying `status: speculative` assertions, merged as `proven` when confirmed or — deliberately — merged as `disproven` when refuted, so negative findings stay on the record), collaboration patterns (shared-repository for family teams, fork-and-PR maintainer model for societies, PR review as genealogical proof-standard review, CI validation), merging and conflict resolution (the `git merge` vs `glx merge` distinction, resolving textual conflicts vs recording evidential conflicts as `status: disputed`, post-merge `glx validate`/`glx duplicates`), history as a research log (`git log --follow`, `git blame`, milestone tags, correct-don't-erase), and scale concerns (Git LFS for `media/files/`, partial clone). Registered in the website nav and guides sidebar, and cross-linked from the Quickstart, Best Practices, and Hands-On CLI guides. Closes #508.
 - **Deterministic memory-file drift detection (`make check-memory-drift`)** — A new offline, API-key-free Go tool under `tools/memcheck` checks every `CLAUDE.md` (and `AGENTS.md` if adopted) against the repository and fails when an instruction has rotted: it asserts that every referenced `make` target exists in the `Makefile`, every concrete file/directory path resolves on disk (relative to the repo root or the memory file's own directory), and every `github.com/<org>/<repo>` import path under the module's org matches the `go.mod` module line. It closes the project's own dogfooding gap (it lints docs↔schema parity in CI but not memory-file↔reality parity) and runs as a dedicated CI job in `validate-spec.yml` alongside `check-code-drift`. Path references are recognized syntactically (a slash plus a known file extension, or a multi-segment trailing-slash directory) so look-alikes such as `actions/checkout`, `os/exec`, and the `claude/` branch prefix don't produce false positives; intentional exceptions live in `.claude/memory-drift-allowlist.yaml`. Surfaced and corrected two stale fixture paths in the root `CLAUDE.md` in the process. (#1053, epic #1016)
