+++ glx/media_copy_test.go
@@ -24,80 +24,6 @@ import (
 	glxlib "github.com/genealogix/glx/go-glx"
 )
 
-func TestDecodeGEDCOMBlob(t *testing.T) {
-	// GEDCOM BLOB encoding: each char minus '.' (0x2E) gives 6 bits.
-	// 4 chars → 3 bytes.
-	// '.' = 0, '/' = 1, '0' = 2, etc.
-
-	// Simple test: 4 dots should produce 3 zero bytes
-	decoded, err := decodeGEDCOMBlob("....")
-	if err != nil {
-		t.Fatalf("Decode failed: %v", err)
-	}
-	if len(decoded) != 3 {
-		t.Fatalf("Expected 3 bytes, got %d", len(decoded))
-	}
-	for i, b := range decoded {
-		if b != 0 {
-			t.Errorf("Byte %d: expected 0, got %d", i, b)
-		}
-	}
-
-	// Test with newlines/whitespace (should be stripped)
-	decoded2, err := decodeGEDCOMBlob("....\n....")
-	if err != nil {
-		t.Fatalf("Decode with newlines failed: %v", err)
-	}
-	if len(decoded2) != 6 {
-		t.Fatalf("Expected 6 bytes, got %d", len(decoded2))
-	}
-
-	// Empty blob
-	_, err = decodeGEDCOMBlob("")
-	if err == nil {
-		t.Error("Expected error for empty blob")
-	}
-
-	// Trailing 3 chars (valid — produces 2 bytes)
-	decoded3, err := decodeGEDCOMBlob("...")
-	if err != nil {
-		t.Fatalf("Decode of 3 chars failed: %v", err)
-	}
-	if len(decoded3) != 2 {
-		t.Errorf("Expected 2 bytes from 3 chars, got %d", len(decoded3))
-	}
-
-	// Trailing 2 chars (valid — produces 1 byte)
-	decoded4, err := decodeGEDCOMBlob("..")
-	if err != nil {
-		t.Fatalf("Decode of 2 chars failed: %v", err)
-	}
-	if len(decoded4) != 1 {
-		t.Errorf("Expected 1 byte from 2 chars, got %d", len(decoded4))
-	}
-
-	// Mixed: 4 + 2 trailing chars (produces 3 + 1 = 4 bytes)
-	decoded5, err := decodeGEDCOMBlob("......")
-	if err != nil {
-		t.Fatalf("Decode of 6 chars failed: %v", err)
-	}
-	if len(decoded5) != 4 {
-		t.Errorf("Expected 4 bytes from 6 chars, got %d", len(decoded5))
-	}
-
-	// Single char is invalid (not enough bits for a full byte)
-	_, err = decodeGEDCOMBlob(".")
-	if err == nil {
-		t.Error("Expected error for single character")
-	}
-
-	// 5 chars (4 + 1 trailing) is invalid — trailing single char can't encode a byte
-	_, err = decodeGEDCOMBlob(".....")
-	if err == nil {
-		t.Error("Expected error for 5 chars (trailing single character)")
-	}
-}
-
 func TestCopyMediaFiles_FileCopy(t *testing.T) {
 	// Set up source directory with a test file
 	srcDir := t.TempDir()
